Added commenting.
Fixed the process of what is an IP Block by stating only subnets that cant have IP addresses assigned are IP Blocks to ensure smooth operation and to prevvent having to come up with a custom solution for IP assignment, IP Range allocation and de-allocation.

This process does mean that the integration for on-demand networks have a manual base requirement e.g. carving up of what will be a ip block and what will be a IP Range.